CRS and SRS review
i removed the old folder of the CRS and removes the old folder of the SIQ
i merged them in one folder which contains the old SIQ , the old CRS and the new excel of both of them together after review.

i added the notification page flowchart, adjusted the add content flowchart and added the categories pages flowchart.